rename a few camelCase stores to kebab-case
