+++ docs/architecture/NAMING.md
@@ -2,7 +2,7 @@
 
 ## Files
 
-- `src/stores/` holds in-memory Solid reactive state — name files `<x>Store.ts` (`contactsStore.ts`, `filesStore.ts`).
+- `src/stores/` holds in-memory Solid reactive state — name files `<x>-store.ts` (`contacts-store.ts`, `files-store.ts`).
 - `src/storage/` holds persistence — name files `<x>-repository` / `<x>-adapter`, never "store".
 - In-memory module state observed via the event bus → `<module>-state.js`.
 - Module barrel: `src/<module>/index.{js,ts}`.
