+++ docs/architecture/STRUCTURE.md
@@ -13,7 +13,7 @@ their direction of allowed imports:
 | `src/infra/`        | External-SDK wiring: Firebase init, App Check, RTDB client, Sentry.                                                                             | `infra`, `lib`                                                                  |
 | `src/storage/`      | Durable persistence ports (`contacts/`, `conversations/`, `files/`, `user/`). D1 + R2 spine; RTDB remains for user presence and legacy cleanup. | `storage`, `shared`, `lib`, `infra`                                             |
 | `src/realtime/`     | Ephemeral coordination (WebRTC signaling, conversation channel, user mailbox). Backed by Cloudflare Durable Objects in `backend/cloudflare/`.   | `realtime`, `shared`, `lib`, `infra`, `auth`                                    |
-| `src/stores/`       | Solid reactive mirrors over storage/realtime (`contactsStore`, `filesStore`, `selectedConversationStore`, …).                                   | `stores`, `auth`, `shared`, `lib`, `storage`, `realtime`, `infra`, `feature`    |
+| `src/stores/`       | Solid reactive mirrors over storage/realtime (`contacts-store`, `files-store`, `conversation-store`, ...).                                      | `stores`, `auth`, `shared`, `lib`, `storage`, `realtime`, `infra`, `feature`    |
 | `src/auth/`         | Session/auth state and helpers.                                                                                                                 | `auth`, `shared`, `lib`, `infra`, `components`                                  |
 | `src/features/<x>/` | Feature modules. Each has a barrel, optional `components/`, and optional `setup()`.                                                             | `auth`, `shared`, `lib`, `feature`, `components`, `infra`, `stores`, `realtime` |
 | `src/components/`   | App-level + shared/primitive UI (`app/`, `base-legacy/`, `dialogs/`, `media/`). **Not** feature UI.                                             | `components`, `auth`, `shared`, `lib`                                           |
